dict_register: Safer multiple registration support
Change DICT.reg_name to store the table's original name, not the most
recently registered name.  (The original name is the name generated by
dict_make_registered_name() in dict_open().)  This fixes a leak of the
previous DICT.reg_name when registering a DICT multiple times.

Also, when registering a DICT under a new name (alias), increment the
refcount associated with the original name.  This increment is removed
once the refcount on the new alias drops to zero (the hash table node
for the alias "owns" the refcount increment added to the original
name).  This fixes a longstanding use-after-free bug where dropping
the refcount on any registered name to zero would invalidate any
references held under different names for the same table.  (This also
avoids a double-free if any other name's refcount also drops to zero.)

Also, do not move DICT.close to DICT.saved_close for the second and
subsequent registrations.  This fixes a panic if a table is registered
multiple times and the refcount for one of the names drops to
zero.  (Previously, both dict->close and dict->saved_close would be
set to dict_register_close(), causing an attempted unregistration on a
name whose refcount is already zero.)

Also add tests.