test_support: Add FAIL_STOP, ERROR to TEST_RESULT enum
This makes it possible for a test case to abort a test run.

FAIL_STOP differs from FAIL in that it causes the subsequent test
cases to not run.  FAIL_STOP is intended to indicate that subsequent
test cases are not meaningful or well-behaved due to the failure so
they should not even run.

ERROR differs from FAIL_STOP in that it is propagated to the caller of
run_tests().  ERROR is intended to indicate a problem with the test
code itself, not the code under test.

This change inverts the zero/non-zero semantics of TEST_RESULT.  Now
zero means PASS, and non-zero means a non-PASS condition.  This
reduces the likelihood of misinterpreting a value that does not match
any of the enum constants (e.g., `return -1;`) as if it was a PASS.
It also means that a TEST_RESULT can be directly used as the test
utility's exit status.